+++ XIVAuras/Config/LabelStyleConfig.cs
@@ -34,7 +34,7 @@ namespace XIVAuras.Config
 
         public IConfigPage GetDefault() => new LabelStyleConfig("[value]");
 
-        public void DrawConfig(Vector2 size, float padX, float padY)
+        public void DrawConfig(IConfigurable parent, Vector2 size, float padX, float padY)
         {
             if (ImGui.BeginChild("##LabelStyleConfig", new Vector2(size.X, size.Y), true))
             {