+++ XIVAuras/Config/StatusTrigger.cs
@@ -53,7 +53,7 @@ namespace XIVAuras.Config
                 data.Value = 10;
                 data.Stacks = 2;
                 data.MaxStacks = 2;
-                data.Icon = this.TriggerData.FirstOrDefault().Icon;
+                data.Icon = this.TriggerData.FirstOrDefault()?.Icon ?? 0;
                 return true;
             }
 
@@ -63,36 +63,22 @@ namespace XIVAuras.Config
                 return false;
             }
 
-            GameObject? actor = this.TriggerSource switch
-            {
-                TriggerSource.Player => player,
-                TriggerSource.Target => Utils.FindTarget(),
-                TriggerSource.TargetOfTarget => Utils.FindTargetOfTarget(),
-                TriggerSource.FocusTarget => Singletons.Get<TargetManager>().FocusTarget,
-                _ => null
-            };
-
-            if (actor is not BattleChara chara)
-            {
-                return false;
-            }
-
             bool active = false;
-            data.Icon = this.TriggerData.FirstOrDefault().Icon;
-            foreach (TriggerData trigger in this.TriggerData)
+            StatusHelpers helper = Singletons.Get<StatusHelpers>();
+            foreach(TriggerData trigger in this.TriggerData)
             {
-                foreach (var status in chara.StatusList)
+                foreach (var status in helper.GetStatus(this.Source, trigger.Id))
                 {
                     if (status is not null &&
-                        status.StatusId == trigger.Id &&
                         (status.SourceID == player.ObjectId || !this.OnlyMine))
                     {
                         active = true;
-                        data.Id = trigger.Id;
+                        data.Id = status.StatusId;
                         data.Value = Math.Abs(status.RemainingTime);
                         data.Stacks = status.StackCount;
                         data.MaxStacks = trigger.MaxStacks;
                         data.Icon = trigger.Icon;
+                        break;
                     }
                 }
             }
@@ -124,7 +110,7 @@ namespace XIVAuras.Config
                 this.TriggerData.Clear();
                 if (!string.IsNullOrEmpty(_triggerNameInput))
                 {
-                    SpellHelpers.FindStatusEntries(_triggerNameInput).ForEach(t => AddTriggerData(t));
+                    StatusHelpers.FindStatusEntries(_triggerNameInput).ForEach(t => AddTriggerData(t));
                 }
 
                 _triggerNameInput = this.TriggerName;