+++ XIVAuras/Helpers/ConfigHelpers.cs
@@ -23,6 +23,16 @@ namespace XIVAuras.Helpers
             SerializationBinder = new XIVAurasSerializationBinder()
         };
 
+        public static T? SerializedClone<T>(T? obj) where T: class?
+        {
+            if (obj == null)
+            {
+                return null;
+            }
+
+            return JsonConvert.DeserializeObject<T>(JsonConvert.SerializeObject(obj, Formatting.None, _serializerSettings), _serializerSettings);
+        }
+
         public static void ExportToClipboard<T>(T toExport)
         {
             string? exportString = ConfigHelpers.GetExportString(toExport);
@@ -241,6 +251,7 @@ namespace XIVAuras.Helpers
             typeof(FontConfig),
             typeof(FontData),
             typeof(IconStyleConfig),
+            typeof(ItemCooldownTrigger),
             typeof(LabelStyleConfig),
             typeof(StatusTrigger),
             typeof(TriggerConfig),